service: extract FrontPage from Paperback
